Move datatype module one level up, makes more sense
