+++ src/lgdo/lh5/_serializers/vector_of_vectors.py
@@ -10,8 +10,8 @@ from ...types import (
     Array,
     VectorOfVectors,
 )
+from .. import datatype as dtypeutils
 from ..exceptions import LH5DecodeError
-from . import datatype as dtypeutils
 from .array import (
     _h5_read_array,
 )
